Pre processing of data to create data base
The results of each race is stored online as a PDF. The preprocessing of the data has two stages - first taking the data from a PDF on th eUCI website and formatting it into a CSV file using excel. Then using Pandas to preprocess the data so it is ready for analysis. Initially I tried to use Numpy to process the data, but Pandas was a much better option.