Improved approach to data collection
Previously I used excel to copy and paste the race time data from a PDF to then sort it into a suitable CSV format to then import it using panadas and clean the data. To build a larger data set with multiple race times in this approach was not suitable. Using tabula-py I can instead read the tables in the PDF directly into a Pandas data frame and then pre-process the data. 

First commit covers reading the table from the PDF and then re-arranging the DataFrame into a suitable format.