Preprocessing function completed
The function will now clean and combine the tables from the pdf into one DataFrame ready for further analysis. The issue this has now is that the many other race time information is stored in other PDF. When using tabula to read these PDFs it doesn't always correctly read the tables. The idea was to write one function that would clean the extracted tables from the PDFs - the extracted data from tabula needs to be in a consistent format for the function to work. I need to find another way to read the data from the PDFs for this to work.